Move RichText component into its own folder
