+++ mathesar_ui/src/components/rich-text/index.ts
@@ -1,0 +1,1 @@
+export { default as RichText } from './RichText.svelte';
