+++ mathesar_ui/src/systems/table-view/link-table/LinkTypeOption.svelte
@@ -4,7 +4,7 @@
   import type { FieldStore } from '@mathesar/components/form';
   import { assertExhaustive } from '@mathesar/utils/typeUtils';
   import { LL } from '@mathesar/i18n/i18n-svelte';
-  import RichText from '@mathesar/components/RichText.svelte';
+  import { RichText } from '@mathesar/components/rich-text';
   import Diagram from './diagram/Diagram.svelte';
   import Pill from './LinkTablePill.svelte';
   import type { LinkType } from './linkTableUtils';
